hexagonal/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
psycopg[binary]==3.2.3
python-dotenv==1.0.1
pytest==8.3.3
